Updated project for iOS8

+++ NGVaryingGridView/NGVaryingGridView/NGVaryingGridView.m
@@ -391,7 +391,7 @@
     }
 }
 
-- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(float)scale {
+- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(CGFloat)scale {
     if ([self.scrollViewDelegate respondsToSelector:@selector(scrollViewDidEndZooming:withView:atScale:)]) {
         [self.scrollViewDelegate scrollViewDidEndZooming:scrollView withView:view atScale:scale]; 
     }
